Card component before adding the 'Add to Cart' button
